check {rhub}	R Documentation

Check an R package on r-hub

Description

Check an R package on r-hub

Usage

check(path = ".", platform = NULL, email = NULL, valgrind = FALSE,
  check_args = character(), env_vars = character(),
  show_status = interactive())

Arguments

path
Path to a directory containing an R package, or path to source R package tarball built with R CMD build or devtools::build().

platform
Platform to build/check the package on. See platforms() for the available platforms. If this is NULL, and the R session is interactive, then a menu is shown. If it is NULL, and the session is not interactive, then the default r-hub platforms is used. Can take a vector of platforms which saves time by building one R package tarball that is used for all the platforms specified.

email
Email address to send notification to about the build. It must be a validated email address, see validate_email(). If NULL, then the email address of the maintainer is used, as defined in the DESCRIPTION file of the package.

valgrind
Whether to run the check in valgrind. Only supported on Linux currently, and ignored on other platforms.

check_args
Extra arguments for the R CMD check command.

env_vars
Environment variables to set on the builder machine before the check. A named character vector.

show_status
Whether to show the status of the build as it is happening.

Value

An rhub_check object.

Examples

## Not run: 
check(".")
check("mypackage_1.0.0.tar.gz", platform = "fedora-clang-devel")

## End(Not run)

[Package rhub version 1.0.2 Index]