rc_list_repos {rhub}	R Documentation

List your repositories created by the R Consortium runners

Description

Lists repositories created by rc_submit() submissions.

Usage

rc_list_repos(email = NULL)

Arguments

email
Email address. We try to detect this, but if the detection fails, you can specify it explicitly.

Value

Data frame with columns:

repo_name: Name of the repository.

repo_url: URL of the repository.

builds_url: URL to the builds of the repository.

Additional columns and customized printing will be probably added later to the result.

See Also

Other RC runners API: rc_list_local_tokens(), rc_new_token(), rc_submit()

[Package rhub version 2.0.1 Index]