rc_new_token {rhub}	R Documentation

Request a new token for submissions to the R Consortium runners

Description

To build and check R packages on the RC runners of R-hub, you'll need to verify your email address. R-hub will send a token to your email address, and this token will be stored on your computer.

Usage

rc_new_token(email = NULL, token = NULL)

Arguments

email
Email address to verify We try to detect this, but if the detection fails, you can specify it explicitly. If this argument is missing (or NULL), then you can specify it interactively.

token
Token to add. If you already received a token in an email from R-hub, you can specify that here.

Details

You need to store a token on every computer you want to submit jobs from, either using the same token from the email you got, or you can request additional tokens for the new machines. Your old token will stay valid as well.

If you already have a token from a previous version of R-hub, you can reuse that and you don't need to do anything.

Run

rhub:::email_file()

to see the file rhub uses to store your tokens.

See Also

Other RC runners API: rc_list_local_tokens(), rc_list_repos(), rc_submit()

[Package rhub version 2.0.1 Index]