rc_submit {rhub}	R Documentation

Submit a package to the R Consortium runners

Description

Submit a package to the R Consortium runners

Usage

rc_submit(path = ".", platforms = NULL, email = NULL, confirmation = NULL)

Arguments

path
Path to package file or package directory.

platforms
Platforms to checks. See rhub_platforms() for a current list. If not specified, then you can select the platforms interactively. Must be specified in non-interactive sessions.

email
Email address. You must have a token on the local machhine, that corresponds to the email address, see rc_new_token(). If not specified (or NULL) then the email address of the package maintainer is used.

confirmation
You must set this to TRUE to submit a package from a non-interactive session.

Value

A list with data about the submission, invisibly. Currently it has:

result: Should be the string "OK".

repo_url: URL to the repository.

actions_url: URL to the builds inside the repository.

id: Build id. This is a string with a randomly generated id.

name: Build name, this is a string, the concatenation of the build platforms.

More fields might be added later.

See Also

rhub_platforms() for a list of supported platforms.

Other RC runners API: rc_list_local_tokens(), rc_list_repos(), rc_new_token()

[Package rhub version 2.0.1 Index]