rhub_check {rhub}	R Documentation

Check a package on R-hub

Description

Check a package on R-hub

Usage

rhub_check(gh_url = NULL, platforms = NULL, r_versions = NULL, branch = NULL)

Arguments

gh_url
GitHub URL of a package to check, or NULL to check the package in the current directory.

platforms
Platforms to use, a character vector. Use NULL to select from a list in interactive sessions. See rhub_platforms().

r_versions
Which R version(s) to use for the platforms that supports multiple R versions. This argument is not implemented yet.

branch
Branch to use to run R-hub. Defaults to the current branch if gh_url is NULL. Otherwise defaults to "main". Note that this branch also need to include the rhub.yaml workflow file.

Value

TODO

[Package rhub version 2.0.1 Index]