rhub_doctor {rhub}	R Documentation

Check if the current or the specified package is ready to use with R-hub

Description

Errors if the package or repository is not set up correctly, and advises on possible solutions.

Usage

rhub_doctor(gh_url = NULL)

Arguments

gh_url
Use NULL for the package in the current working directory. Alternatively, use the URL of a GitHub repository that contains an R package that was set up to use with R-hub.

[Package rhub version 2.0.1 Index]