rhub_platforms {rhub}	R Documentation

List R-hub platforms

Description

List R-hub platforms

Usage

rhub_platforms()

Value

Data frame with columns:

name: platform name. Use this in the platforms argument of rhub_check().

aliases: alternative platform names. They can also be used in the platforms argument of rhub_check().

type: "os" or "container".

os_type: Linux, macOS or Windows currently.

container: URL of the container image for container platforms.

github_os: name of the OS on GitHub Actions for non-container platforms.

r_version: R version string. If "*" then any supported R version can be selected for this platform.

os_name: name of the operating system, including Linux distribution name and version for container actions.

[Package rhub version 2.0.1 Index]