rhub_setup {rhub}	R Documentation

Setup the current R package for use with R-hub

Description

It adds or updates the R-hub workflow file to the current package, and advises on next steps.

Usage

rhub_setup(overwrite = FALSE)

Arguments

overwrite
if TRUE, rhub_setup() will overwrite an already existing workflow file.

Value

Nothing.

[Package rhub version 2.0.1 Index]