validate_email {rhub}	R Documentation

Validate an email address on r-hub

Description

To build and check R packages on r-hub, you need to validate your email address. This is because r-hub sends out emails about build results.

Usage

validate_email(email = NULL, token = NULL)

Arguments

email
The email address to validate.

token
Token obtained from rhub, to validate the email address.

Details

The rhub package stores validated email addresses in a user configuration file, at a platform-dependent location. On your current platform the file is at /home/ashwini.jeggari/.local/share/rhub/validated_emails.csv.

To validate a new email address, call this function from an interactive R session, without any arguments.

To add an email address that was validated before (probably on another machine), to the configuration file, call this function with the email and token arguments.

See Also

Other email validation: list_validated_emails

[Package rhub version 1.0.2 Index]