check_on_linux {rhub}	R Documentation

Check an R package on an r-hub platform

Description

These functions provide a quick easy to use interface to check a package on a platform with some particular aspect. Which platform they use might change over time.

Usage

check_on_linux(path = ".", ...)

check_on_windows(path = ".", ...)

check_on_macos(path = ".", ...)

check_on_debian(path = ".", ...)

check_on_ubuntu(path = ".", ...)

check_on_fedora(path = ".", ...)

check_on_centos(path = ".", ...)

check_with_roldrel(path = ".", ...)

check_with_rrelease(path = ".", ...)

check_with_rpatched(path = ".", ...)

check_with_rdevel(path = ".", ...)

check_with_valgrind(path = ".", ...)

check_with_sanitizers(path = ".", ...)

Arguments

path
Path to a directory containing an R package, or path to source R package tarball built with R CMD build or devtools::build().

...
Additional arguments are passed to check().

Value

An rhub_check object.

[Package rhub version 1.0.2 Index]