An rhub_check object holds status and results of rhub checks

This class is similar to rhub_check (it inherits from it), but it shows the results of the checks in a more concise format by default. To show the details, call the $details() method.

Usage

ch$update()
ch$print(...)
ch$web(which = NULL)
ch$livelog(which = 1)

chl$update()
chl$print(...)
chl$details()
chl$livelog(which = 1)

Arguments

ch An rhub check object. It can be created using check(), list_my_checks() or list_package_checks(). See also last_check().

... Extra arguments are currently ignored.

which Which check to show, if the object contains multiple checks. For web the default is all checks. For livelog the default is the first check. A check can be selected via its number or id.

Details

An rhub_check object can be created by check(), list_my_checks(), or list_package_checks(). last_check() returns the last check(s) submitted from the current R session.

ch$update() updates the status of the check. Printing the check status to the screen does not perform an update, unless the status of the check(s) is unknown.

ch$print() prints the status of the check(s) to the screen.

ch$web() opens a tab or window in the default web browser, that points to the detailed logs of the check.

ch$livelog() shows the live log of the check. The live log can be interrupted using the usual interruption keyboard shortcut, usually CTRL+c or ESC.

The list_my_checks() and list_package_checks() functions create rhub_check_list objects.

See rhub_check for the description of the inherited chl$update(), chl$print(), chl$livelog() methods.

chl$details() prints the details for all checks in the list.

Examples

# NOT RUN {
check()
ch <- last_check()
ch$update()
ch$web()
ch$livelog()
# }# NOT RUN {
chl <- list_my_checks()
chl
chl$details()
# }