An rhub_check object holds status and results of rhub checks

An rhub_check object holds status and results of rhub checks

Usage

ch <- rhub_check$new(ids = NULL, status = NULL, group = NULL)
ch$get_ids()
ch$update()
ch$print(...)
ch$browse(which = NULL)
ch$urls(which = NULL)
ch$livelog(which = 1)
ch$cran_summary()

Arguments

ch An rhub check object. It can be created using check(), and other check functions including check_for_cran. See also last_check().

ids Character vector of check ids.

status Check status for ids or group.

group Check group id, string scalar. Either group or ids must be non-NULL.

... Extra arguments are currently ignored.

which Which check to show, if the object contains multiple checks. For browse the default is all checks. For livelog the default is the first check. A check can be selected via its number or id.

Details

An rhub_check object can be created by check(), list_my_checks(), or list_package_checks(). last_check() returns the last check(s) submitted from the current R session. Do not confuse rhub_check/rhub_check_for_cran (classes) with check() or check_for_cran() (functions).

ch$get_ids() returns the check ids. These can be used to query if a check has finished.

ch$update() updates the status of the check. Printing the check status to the screen does not perform an update, unless the status of the check(s) is unknown.

ch$print() prints the status of the check(s) to the screen.

ch$cran_summary() prints text to be copy-pasted in cran-comments.md, it is especially useful on the output of check_for_cran().

ch$browse() opens a tab or window in the default web browser, that points to the detailed logs of the check(s).

ch$urls() return a tibble::tibble with URL to the html log, text log and artifacts of the check(s).

For both ch$browse() and ch$urls(), note that the logs and artifacts are not kept forever, they are accessible for a few days after submission.

ch$livelog() shows the live log of the check. The live log can be interrupted using the usual interruption keyboard shortcut, usually CTRL+c or ESC.